backend/requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-cors==5.0.0
werkzeug==3.1.3
python-dotenv==1.0.1
